requirements.txt
fastapi
uvicorn
paho-mqtt
msgpack-python
minio
pydantic-settings

#Test only
pytest
pytest-docker